Move easter egg player file to utils
