+++ lib/utils/easter_egg_player.dart
